requirements.txt
importlib-metadata==4.0.1
lmdb==1.2.1
matplotlib==3.3.4
numpy==1.18.5
nvidia-cublas==11.2.1.74
nvidia-cuda-cupti==11.1.69
nvidia-cuda-nvcc==11.1.74
nvidia-cuda-nvrtc==11.1.74
nvidia-cuda-runtime==11.1.74
nvidia-cudnn==8.0.4.30
nvidia-cufft==10.3.0.74
nvidia-curand==10.2.2.74
nvidia-cusolver==11.0.0.74
nvidia-cusparse==11.2.0.275
nvidia-dali-cuda110==0.26.0
nvidia-dali-nvtf-plugin==0.26.0+nv20.10
nvidia-nccl==2.7.8
nvidia-tensorboard==1.15.0+nv20.10
nvidia-tensorflow==1.15.4+nv20.10
nvidia-tensorrt==7.2.1.4
packaging==21.0
pathlib2
Pillow==8.2.0
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-shutil==1.7.0
python-dateutil==2.8.1
PyYAML==5.4.1
scikit-learn==0.24.2
scipy==1.5.4
sklearn==0.0
tensorflow-estimator==1.15.1
tensorflow-model-optimization==0.6.0
termcolor==1.1.0
threadpoolctl==3.0.0
toml==0.10.2
tqdm==4.62.2
typing-extensions==3.10.0.0
urllib3==1.26.6
requests==2.27.1